dalvikvm fixes (previously wrongly initialized), FruitNinja now starts up with --use-dvm
